Update Cage CLI to use relative paths for key and cert

+++ src/cert/mod.rs
@@ -3,13 +3,13 @@ use itertools::Itertools;
 use rcgen::CertificateParams;
 use std::io::Write;
 use std::ops::Add;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 pub mod error;
 pub use error::CertError;
 
 pub fn create_new_cert(
-    output_dir: &str,
+    output_dir: &Path,
     distinguished_name: DistinguishedName,
 ) -> Result<(PathBuf, PathBuf), CertError> {
     let mut cert_params = CertificateParams::new(vec![]);
@@ -61,20 +61,19 @@ fn add_distinguished_name_to_cert_params(
 }
 
 fn write_cert_to_fs(
-    output_path: &str,
+    output_path: &Path,
     cert: rcgen::Certificate,
-) -> Result<(std::path::PathBuf, std::path::PathBuf), CertError> {
-    let output_path = std::path::Path::new(output_path);
-    let path = output_path
-        .canonicalize()
-        .map_err(|_| CertError::OutputPathDoesNotExist)?;
+) -> Result<(PathBuf, PathBuf), CertError> {
+    if !output_path.exists() {
+        return Err(CertError::OutputPathDoesNotExist);
+    }
 
-    let cert_path = path.join("cert.pem");
+    let cert_path = output_path.join("cert.pem");
     let mut cert_file = std::fs::File::create(cert_path.as_path())?;
     let serialized_cert = cert.serialize_pem()?;
     cert_file.write_all(serialized_cert.as_bytes())?;
 
-    let key_path = path.join("key.pem");
+    let key_path = output_path.join("key.pem");
     let mut key_file = std::fs::File::create(key_path.as_path())?;
     let serialized_key = cert.serialize_private_key_pem();
     key_file.write_all(serialized_key.as_bytes())?;
